django-web/webapp/requirements.txt
django
djangorestframework
django-cors-headers
python-dotenv
transformers
sentence-transformers
faiss-cpu
langchain
openai
